Migrate chain queries out of Node.py into their own class
Node inherits from new Queries class to avoid intrusive changes to tests
processCurlCommand is left in Node to avoid merge conflicts with #442
