+++ tests/TestHarness/CMakeLists.txt
@@ -4,3 +4,4 @@ configure_file(WalletMgr.py WalletMgr.py COPYONLY)
 configure_file(Node.py Node.py COPYONLY)
 configure_file(Cluster.py Cluster.py COPYONLY)
 configure_file(TestHelper.py TestHelper.py COPYONLY)
+configure_file(queries.py queries.py COPYONLY)
